review both controllers country& discount

+++ PropertyRentingApi/Controllers/DiscountController.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using PR_BusinessLayer;
-using PR_DataAccessLayer;
+using SharedDTOLayer.Offer.DiscountDTO;
+using System.Security.Claims;
 
 namespace PropertyRentingApi.Controllers
 {
@@ -9,7 +9,7 @@ namespace PropertyRentingApi.Controllers
     [ApiController]
     public class DiscountController : ControllerBase
     {
-        [HttpGet("getDicountbyID", Name = "GetDiscountById")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("getDicountbyID", Name = "GetDiscountById")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -34,7 +34,7 @@ namespace PropertyRentingApi.Controllers
             return Ok(DDTO);
         }
 
-        [HttpGet("getActiveDicountbyPropertyID", Name = "getActiveDicountbyPropertyID")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("getActiveDicountbyPropertyID", Name = "getActiveDicountbyPropertyID")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -49,18 +49,19 @@ namespace PropertyRentingApi.Controllers
             }
 
             int ActiveDiscount = PR_BusinessLayer.clsProperty.FindActiveDiscountByID(PropertyID);
+            
             if (ActiveDiscount == -1)
             {
-                return NotFound($"Property with ID {PropertyID} not found or No Valid Offer is exist for this Property.");
+                return NotFound($"There's no active offer with propertyID= {PropertyID}");
             }
-
+            
             
 
             return Ok(ActiveDiscount);
         }
 
 
-        [HttpGet("GetAllPropertyDicounts", Name = "GetDiscountsByPropertyId")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("GetAllPropertyDicounts", Name = "GetDiscountsByPropertyId")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -77,7 +78,7 @@ namespace PropertyRentingApi.Controllers
             List<DiscountDTO> discounts = PR_BusinessLayer.clsDiscount.GetAllDiscountsByPropertyID(PropertyID);
             if (discounts == null)
             {
-                return NotFound($"Property with ID {PropertyID} not found.");
+                return NotFound($"Offers with property  {PropertyID}= ID are empty.");
             }
 
             return Ok(discounts);
@@ -85,7 +86,7 @@ namespace PropertyRentingApi.Controllers
 
 
 
-        [HttpGet("GetAllDiscountsBelongToAproperty", Name = "GetAllDiscountsBelongToAproperty")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("GetAllDiscountsBelongToAproperty", Name = "GetAllDiscountsBelongToAproperty")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -102,7 +103,7 @@ namespace PropertyRentingApi.Controllers
             List<PropertyDiscountDTO> discounts = PR_BusinessLayer.clsDiscount.FindAllDiscountBelongToAPropertyByPropertyID(PropertyID);
             if (discounts == null)
             {
-                return NotFound($"Property with ID {PropertyID} not found.");
+                return NotFound($"No Offers are found  for this {PropertyID} !");
             }
 
             return Ok(discounts);
@@ -110,38 +111,7 @@ namespace PropertyRentingApi.Controllers
 
 
 
-
-
-
-        [HttpPost("AddDiscount", Name = "AddDiscount")]
-        [ProducesResponseType(StatusCodes.Status201Created)]
-        [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public ActionResult<DiscountDTO> AddDiscount(DiscountDTO DiscountDTO)
-        {
-            //we validate the data here
-            if (DiscountDTO == null || DiscountDTO.PropertyID == -1 ||
-                DiscountDTO.DiscountPercentage == 0 || DiscountDTO.StartDate == default(DateTime) || DiscountDTO.EndDate == default(DateTime))
-            {
-                return BadRequest("Invalid Discount data.");
-            }
-
-           
-            PR_BusinessLayer.clsDiscount Discount = new PR_BusinessLayer.clsDiscount(new DiscountDTO(DiscountDTO.DiscountID, DiscountDTO.PropertyID,
-                DiscountDTO.DiscountPercentage, DiscountDTO.StartDate, DiscountDTO.EndDate, DiscountDTO.IsDeleted));
-            Discount.Save();
-
-            DiscountDTO.DiscountID = Discount.DiscountID;
-
-          
-
-            return CreatedAtRoute("AddDiscount", new { id = DiscountDTO.DiscountID }, DiscountDTO);
-
-
-        }
-
-
-
-        [HttpGet("GetDiscountID", Name = "GetDiscountID")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("GetDiscountID", Name = "GetDiscountID")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -158,55 +128,76 @@ namespace PropertyRentingApi.Controllers
             int GetDiscountID = PR_BusinessLayer.clsDiscount.Find(DiscountID).DiscountID;
             if (GetDiscountID == 0)
             {
-                return NotFound($"Property with ID {GetDiscountID} not found.");
+                return NotFound($"Offer with ID {GetDiscountID} not found.");
             }
 
             return Ok(GetDiscountID);
         }
 
 
-
+        [Authorize] 
         [HttpPost("AddDiscountToProperty", Name = "AddDiscountToProperty")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         public ActionResult<DiscountDTO> AddDiscountToProperty(int PropertyID, decimal PrecentOff, string StartDate, string EndDate)
         {
             DiscountDTO DiscountDTO = new DiscountDTO();
-            //we validate the data here
+      
+
+
             if ( PropertyID == -1 ||
                 PrecentOff == 0 || StartDate == "" || EndDate =="")//default(DateTime)
             {
                 return BadRequest("Invalid Discount data.");
             }
 
-        
-            PR_BusinessLayer.clsDiscount Discount =  PR_BusinessLayer.clsProperty.PropertyPricePrecentOFF(PropertyID , PrecentOff , StartDate  ,EndDate);
-            Discount.Save();
 
-            DiscountDTO.DiscountID = Discount.DiscountID;
+              var userIdStr = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+              int PropertyOwner = PR_BusinessLayer.clsPropertyOwner.GetPropertyClientIdByPropertyID(PropertyID);
 
+              if (Convert.ToInt32(userIdStr) != PropertyOwner) return Unauthorized("you're  Unauthorized to add  Offer!");
             
 
-            return CreatedAtRoute("AddDiscount", new { id = DiscountDTO.DiscountID }, DiscountDTO);
+            //start date must set like 2025-08-19
+            PR_BusinessLayer.clsDiscount Discount =  PR_BusinessLayer.clsProperty.PropertyPricePrecentOFF(PropertyID , PrecentOff , StartDate  ,EndDate);
+            
+
+            
+            DiscountDTO.DiscountID = Discount.DiscountID;
+            DiscountDTO.PropertyID = Discount.PropertyID;
+            DiscountDTO.StartDate = Discount.StartDate;
+            DiscountDTO.DiscountPercentage = Discount.DiscountPercentage;
+            DiscountDTO.EndDate = Discount.EndDate;
+            DiscountDTO.IsDeleted = Discount.IsDeleted;
+           
+            return CreatedAtRoute("AddDiscountToProperty", new { id = DiscountDTO.DiscountID }, DiscountDTO);
 
 
         }
 
+
+        [Authorize]
         [HttpPut("UpdateDiscontById", Name = "UpdateDiscount")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public ActionResult<DiscountDTO> UpdateDiscount(int DiscontId, DiscountDTO DiscountDTO)
+        public ActionResult<DiscountDTO> UpdateDiscount( DiscountDTO DiscountDTO)
         {
-            //we validate the data here
+
             if (DiscountDTO == null || DiscountDTO.PropertyID == -1 ||
                 DiscountDTO.DiscountPercentage == 0 || DiscountDTO.StartDate == default(DateTime) || DiscountDTO.EndDate == default(DateTime))
             {
                 return BadRequest("Invalid Discount data.");
             }
 
-          
-            PR_BusinessLayer.clsDiscount Discount = PR_BusinessLayer.clsDiscount.Find(DiscontId);
-          
+
+            var userIdStr = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            int PropertyOwner = PR_BusinessLayer.clsPropertyOwner.GetPropertyClientIdByPropertyID(DiscountDTO.PropertyID);
+
+            if (Convert.ToInt32(userIdStr) != PropertyOwner) return Unauthorized("you're you're  Unauthorized to update this Offer!");
+
+
+            PR_BusinessLayer.clsDiscount Discount = PR_BusinessLayer.clsDiscount.Find(DiscountDTO.DiscountID);
+
 
             Discount.DiscountID = DiscountDTO.DiscountID;
             Discount.PropertyID = DiscountDTO.PropertyID;
@@ -215,14 +206,16 @@ namespace PropertyRentingApi.Controllers
             Discount.EndDate = DiscountDTO.EndDate;
             Discount.Save();
 
-          
+            
+
             return Ok(Discount.DDTO);
 
 
         }
 
 
-        //here we use HttpDelete method
+        
+        [Authorize]
         [HttpDelete("Cancel/{CancelOfferById}", Name = "CancelOfferById")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
@@ -233,13 +226,18 @@ namespace PropertyRentingApi.Controllers
             {
                 return BadRequest($"Not accepted ID {CancelOfferById}");
             }
+            var propertyID = PR_BusinessLayer.clsDiscount.Find(CancelOfferById).PropertyID;
+            var userIdStr = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            int PropertyOwner = PR_BusinessLayer.clsPropertyOwner.GetPropertyClientIdByPropertyID(propertyID);
 
-           
+            if (Convert.ToInt32(userIdStr) != PropertyOwner) return Unauthorized("you're you're  Unauthorized to delete this Offer!");
+
+
             if (PR_BusinessLayer.clsDiscount.DeleteDiscountOffer(CancelOfferById) != -1)
 
-                return Ok($"Property with ID {CancelOfferById} has been deleted.");
+                return Ok($"Offer with ID {CancelOfferById} has been deleted.");
             else
-                return NotFound($"Property with ID {CancelOfferById} not found. no rows deleted!");
+                return NotFound($"Offer with ID {CancelOfferById} not found!");
         }
 
     }
